requirements.txt
# Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.35
PyMySQL==1.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7
werkzeug==3.0.1

# PDF Generation
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
